+++ wallettemplate/src/main/java/wallettemplate/WalletSetPasswordController.java
@@ -23,7 +23,8 @@ import javafx.scene.control.*;
 import javafx.scene.layout.*;
 import org.bitcoinj.crypto.*;
 import org.bitcoinj.wallet.*;
-import org.bitcoinj.walletfx.overlay.OverlayWindowController;
+import org.bitcoinj.walletfx.overlay.OverlayController;
+import org.bitcoinj.walletfx.overlay.OverlayableStackPaneController;
 import org.slf4j.*;
 import org.bouncycastle.crypto.params.*;
 
@@ -35,7 +36,7 @@ import java.util.concurrent.*;
 import org.bitcoinj.walletfx.utils.KeyDerivationTasks;
 import static org.bitcoinj.walletfx.utils.GuiUtils.*;
 
-public class WalletSetPasswordController implements OverlayWindowController<WalletSetPasswordController> {
+public class WalletSetPasswordController implements OverlayController<WalletSetPasswordController> {
     private static final Logger log = LoggerFactory.getLogger(WalletSetPasswordController.class);
     public PasswordField pass1, pass2;
 
@@ -44,7 +45,7 @@ public class WalletSetPasswordController implements OverlayWindowController<Wall
     public Button closeButton;
     public Label explanationLabel;
 
-    private MainController.OverlayUI<? extends OverlayWindowController<WalletSetPasswordController>> overlayUI;
+    private OverlayableStackPaneController.OverlayUI<? extends OverlayController<WalletSetPasswordController>> overlayUI;
     // These params were determined empirically on a top-range (as of 2014) MacBook Pro with native scrypt support,
     // using the scryptenc command line tool from the original scrypt distribution, given a memory limit of 40mb.
     public static final Protos.ScryptParameters SCRYPT_PARAMETERS = Protos.ScryptParameters.newBuilder()
@@ -55,7 +56,7 @@ public class WalletSetPasswordController implements OverlayWindowController<Wall
             .build();
 
     @Override
-    public void setOverlayUI(MainController.OverlayUI<? extends OverlayWindowController<WalletSetPasswordController>> ui) {
+    public void setOverlayUI(OverlayableStackPaneController.OverlayUI<? extends OverlayController<WalletSetPasswordController>> ui) {
         overlayUI = ui;
     }
